+++ libpal/pal_page_allocator.h
@@ -29,9 +29,9 @@ class palPageAllocator : public palAllocatorInterface {
 public:
   palPageAllocator();
 
-  virtual void* Allocate(uint32_t size, int alignment);
+  virtual void* Allocate(uint64_t size, uint32_t alignment);
   virtual void Deallocate(void* ptr);
-  virtual uint32_t GetSize(void* ptr) const;
+  virtual uint64_t GetSize(void* ptr) const;
 
   uint32_t GetPageSize() const;
